+++ lib/cocoapods-core/gem_version.rb
@@ -1,5 +1,5 @@
 module Pod
   # The version of the cocoapods-core.
   #
-  CORE_VERSION = '0.36.2' unless defined? Pod::CORE_VERSION
+  CORE_VERSION = '0.36.3' unless defined? Pod::CORE_VERSION
 end